requirements.txt
Django                        5.0.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.1
pillow                        10.2.0
PyJWT                         2.8.0
PyMySQL                       1.1.0
python-dateutil               2.8.2
requests                      2.31.0